+++ lib/src/interface/interface_scope.dart
@@ -2,6 +2,7 @@ import 'package:cisco_ios_dsl/src/device.dart';
 import 'package:cisco_ios_dsl/src/interface/interface_ether_channel.dart';
 import 'package:cisco_ios_dsl/src/interface/interface_ip.dart';
 import 'package:cisco_ios_dsl/src/interface/interface_ipv6.dart';
+import 'package:cisco_ios_dsl/src/interface/interface_spanning_tree.dart';
 import 'package:cisco_ios_dsl/src/interface/interface_switchport.dart';
 import 'package:cisco_ios_dsl/src/util/toggleable_property.dart';
 
@@ -13,12 +14,14 @@ class InterfaceScope {
   final ToggleableProperty operation;
   final InterfaceSwitchport switchport;
   final InterfaceEtherChannel etherChannel;
+  final InterfaceSpanningTree spanningTree;
 
   InterfaceScope(this.device)
     : ip = InterfaceIp(device),
       ipv6 = InterfaceIpV6(device),
       switchport = InterfaceSwitchport(device),
       etherChannel = InterfaceEtherChannel(device),
+      spanningTree = InterfaceSpanningTree(device),
       operation = ToggleableProperty(device, 'no shutdown', 'shutdown');
 
   void setDescription(String description) => device